requirements.txt
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
click>=8.1.0
jinja2>=3.1.0
